Update README.md
Domain Finder Tool

The Domain Finder is a command-line script designed to discover subdomains for a given domain. It utilizes tools like assetfinder and httprobe to gather and verify subdomains, presenting a sorted list of reachable subdomains. Developed by Aswanth, this tool helps streamline the process of identifying active subdomains, making it useful for security assessments and website management tasks.